🐛 fix(pkrvars.hcl): set vm_disk_thin_provisioned to false The vm_disk_thin_provisioned variable was set to true in all the pkrvars.hcl files. This commit changes the value to false to improve performance by allocating the full disk space upfront instead of dynamically allocating it.
🐛 fix(pkrvars.hcl): set vm_disk_thin_provisioned to false
The vm_disk_thin_provisioned variable was set to true in all the pkrvars.hcl files. This commit changes it to false to ensure that the virtual disks are not thin provisioned. Thin provisioning can cause performance issues and can lead to unexpected behavior when the disk runs out of space.